WIP: Remove Picker's Data type parameter
We can instead capture values from local scopes where we define the
columns. This fixes the borrowing errors.

Switching to dynamic dispatch here is a little meh though. These
functions may be called a lot: potentially on every option and on
every keypress, so it would be nice to do static dispatch.

My current idea for that is to revert this change and move `type = Data`
from the menu::Item trait onto a Column trait and have that basically
do everything menu::Item was doing. It's a little more verbose though
so I'd like to avoid it if possible.