WIP: Support finding the dynamic column for DynPicker
Feels slightly hacky (I wish the type-system helped out here rather
than finding vec indices). The DynPicker can ask the picker the index
of the dynamic column and then use that to fetch the current query
string. This together with the display-only changes works nicely for
the workspace symbol picker. That has been refactored to use a different
set of columns than the document symbol picker, separating the name
and path columns and marking the name column as dynamic.