WIP: Fix matching across all prompts
This could be slightly more ideal if we saved the match state
per-item per-column in a matrix (implemented as a Vec with |I| * |C|
columns). That would allow us to `self.matches.clear()`,
`self.matches.extend()` with all `self.options.iter()` again and not
repeat the work we've already done for the other columns.

Though I'm starting to wonder if we want a lucene-like syntax for
searching individual fields. I think that would be really nice.