WIP: Eliminate unnecessary clone of common highlight byte ranges
It's a little awkward to use the `&mut` like this since the Vecs are
never actually modified, but it's necessary for the sake of the borrow
checker: `highlight_byte_ranges` needs to be a `&Vec` in order for
`common_highlight_byte_ranges` to be shared between loop iterations. We
can't just return a `&Vec` in the else block because it wouldn't live
long enough.